Add automated testing using github actions
it currently builds byteman with debug flags and then tries to parse 112 bitstreams from Xilinx Series 7, UltraScale and UltraScale+ families.
Before parsing the bitstreams, the buffers are set to all 0xFFFFFFFF's. Then at the end we check if any word was not covered by the bitstreams by checking if its value is still 0xFFFFFFFF. This works as none of the 112 bitstreams has a bitstream word with value of 0xFFFFFFFF.
This checks for a full coverage on the device.
The opposite check is if all FAR addresses in the bitstreams are valid, and that check is performed while parsing them in, using assertions.