Use a more generic partName, so device doesn't get reset
It got reset when loading multiple bitstreams or manually loading a device type and there was a difference in the name (e.g., "xcku025" is different than "xcku025-ffva1156-1-c". Now it uses the short name only when looking for device updates. So it would parse "xcku025-ffva1156-1-c" and then it will use "xcku025" as a "shortPartName" for it.